+++ packages/core/database/lib/entity-manager/index.js
@@ -23,10 +23,7 @@ const types = require('../types');
 const { createField } = require('../fields');
 const { createQueryBuilder } = require('../query');
 const { createRepository } = require('./entity-repository');
-const {
-  deleteRelatedMorphOneRelationsAfterMorphToManyUpdate,
-  deleteMorphRelations,
-} = require('./morph-relations');
+const { deleteRelatedMorphOneRelationsAfterMorphToManyUpdate } = require('./morph-relations');
 const {
   isBidirectional,
   isAnyToOne,
@@ -40,6 +37,10 @@ const {
   deleteRelations,
   cleanOrderColumns,
 } = require('./regular-relations');
+const {
+  attachMorphToOneRelation,
+  updateMorphOneToManyRelation,
+} = require('./relations/morph-any-to-many');
 
 const toId = (value) => value.id || value;
 const toIds = (value) => castArray(value || []).map(toId);
@@ -431,7 +432,6 @@ const createEntityManager = (db) => {
 
         if (attribute.relation === 'morphOne' || attribute.relation === 'morphMany') {
           const { target, morphBy } = attribute;
-
           const targetAttribute = db.metadata.get(target).attributes[morphBy];
 
           if (targetAttribute.relation === 'morphToOne') {
@@ -446,51 +446,17 @@ const createEntityManager = (db) => {
               .transacting(trx)
               .execute();
           } else if (targetAttribute.relation === 'morphToMany') {
-            const { joinTable } = targetAttribute;
-            const { joinColumn, morphColumn } = joinTable;
-
-            const { idColumn, typeColumn } = morphColumn;
-
-            const relsToAdd = cleanRelationData.set || cleanRelationData.connect;
-            const relIdsToadd = toIds(relsToAdd);
-
-            if (isEmpty(relIdsToadd)) {
-              continue;
-            }
-
-            // Prepare new relations to insert
-            const insert = relsToAdd.map((data) => {
-              return {
-                [joinColumn.name]: data.id,
-                [idColumn.name]: id,
-                [typeColumn.name]: uid,
-                ...(joinTable.on || {}),
-                ...(data.__pivot || {}),
-                field: attributeName,
-              };
-            });
-
-            // get order value from max value
-            const maxOrder = await db
-              .getConnection()
-              .select(joinColumn.name)
-              .max('order', { as: 'max' })
-              .whereIn(joinColumn.name, relIdsToadd)
-              .where(joinTable.on || {})
-              .groupBy(joinColumn.name)
-              .from(joinTable.name)
-              .transacting(trx);
-
-            const maxMap = maxOrder.reduce((acc, curr) => {
-              acc[curr[joinColumn.name]] = curr.max;
-              return acc;
-            }, {});
-
-            insert.forEach((rel) => {
-              rel.order = (maxMap[rel[joinColumn.name]] || 0) + 1;
-            });
-
-            await this.createQueryBuilder(joinTable.name).insert(insert).transacting(trx).execute();
+            await attachMorphToOneRelation(
+              uid,
+              id,
+              attributeName,
+              targetAttribute,
+              cleanRelationData,
+              {
+                db,
+                trx,
+              }
+            );
           }
 
           continue;
@@ -683,77 +649,17 @@ const createEntityManager = (db) => {
                 .execute();
             }
           } else if (targetAttribute.relation === 'morphToMany') {
-            const { joinTable } = targetAttribute;
-            const { joinColumn, morphColumn } = joinTable;
-
-            const { idColumn, typeColumn } = morphColumn;
-
-            const isPartialUpdate = !has('set', cleanRelationData);
-            const relsToAdd = cleanRelationData.set || cleanRelationData.connect;
-            const relIdsToadd = toIds(relsToAdd);
-
-            if (isPartialUpdate) {
-              // Calculate relations to delete
-              const relIdsToDelete = toIds(
-                differenceWith(isEqual, cleanRelationData.disconnect, relsToAdd)
-              );
-              if (!isEmpty(relIdsToDelete)) {
-                await deleteMorphRelations({
-                  id,
-                  uid,
-                  attributeName,
-                  targetAttribute,
-                  db,
-                  relIdsToDelete,
-                  transaction: trx,
-                });
-              }
-            } else {
-              await deleteMorphRelations({
-                id,
-                uid,
-                attributeName,
-                targetAttribute,
+            await updateMorphOneToManyRelation(
+              uid,
+              id,
+              attributeName,
+              targetAttribute,
+              cleanRelationData,
+              {
+                trx,
                 db,
-                relIdsToDelete: 'all',
-                transaction: trx,
-              });
-            }
-
-            if (isEmpty(relsToAdd)) {
-              continue;
-            }
-
-            const insert = relsToAdd.map((data) => ({
-              [joinColumn.name]: data.id,
-              [idColumn.name]: id,
-              [typeColumn.name]: uid,
-              ...(joinTable.on || {}),
-              ...(data.__pivot || {}),
-              field: attributeName,
-            }));
-
-            // get order value from max value
-            const maxOrder = await db
-              .getConnection()
-              .select(joinColumn.name)
-              .max('order', { as: 'max' })
-              .whereIn(joinColumn.name, relIdsToadd)
-              .where(joinTable.on || {})
-              .groupBy(joinColumn.name)
-              .from(joinTable.name)
-              .transacting(trx);
-
-            const maxMap = maxOrder.reduce((acc, curr) => {
-              acc[curr[joinColumn.name]] = curr.max;
-              return acc;
-            }, {});
-
-            insert.forEach((rel) => {
-              rel.order = (maxMap[rel[joinColumn.name]] || 0) + 1;
-            });
-
-            await this.createQueryBuilder(joinTable.name).insert(insert).transacting(trx).execute();
+              }
+            );
           }
           continue;
         }
